comparepro/requirements.txt
Flask==3.0.3
Flask-Cors==3.0.10
python-dotenv==1.0.1
requests==2.32.3
fuzzywuzzy==0.18.0
python-Levenshtein==0.26.1
rapidfuzz==3.12.1